Refactor character movement: implement smooth movement functions and improve collision handling

+++ include/core.h
@@ -1,8 +1,3 @@
-/**
- * @file core.h
- * @brief Définitions centrales du jeu (constantes, types, structures)
- */
-
 #ifndef NUPRC_CORE_H
 #define NUPRC_CORE_H
 
@@ -15,7 +10,6 @@
 #include <stdbool.h>
 #include <string.h>
 
-/* Configuration de la grille */
 #define GRID_CELL_SIZE      50
 #define GRID_WORLD_WIDTH    256
 #define GRID_WORLD_HEIGHT   88
@@ -27,7 +21,6 @@
 #define MAP_TILES_HEIGHT    8
 #define MAP_TILES_COUNT     (MAP_TILES_WIDTH * MAP_TILES_HEIGHT)
 
-/* Configuration de la fenêtre */
 #define WINDOW_TITLE            "NUPRC - Zelda-like"
 #define WINDOW_FONT_PATH        "assets/fonts/DejaVuSans-Bold.ttf"
 #define WINDOW_FONT_SIZE        16
@@ -35,23 +28,19 @@
 #define WINDOW_WIDTH            (GRID_ROOM_WIDTH * GRID_CELL_SIZE)
 #define WINDOW_HEIGHT           (GRID_ROOM_HEIGHT * GRID_CELL_SIZE + WINDOW_TEXTAREA_HEIGHT)
 
-/* Configuration du gameplay */
 #define GAME_MAX_ENEMIES    10
 #define GAME_INITIAL_LIVES  5
 #define GAME_INITIAL_SCORE  0
 #define GAME_INITIAL_ROOM   {7, 7}
 
-/* Configuration du mouvement fluide */
-#define MOVEMENT_SPEED      0.15f   // Vitesse de déplacement (en cases par frame)
+#define MOVEMENT_SPEED      0.15f
 
-/* Chemins des assets */
 #define ASSET_TEXTURE_LINK      "assets/textures/characters/link2.bmp"
 #define ASSET_TEXTURE_ENEMY     "assets/textures/characters/enemy7.bmp"
 #define ASSET_MAP_TILES         "assets/textures/map/overworldtiles.bmp"
 #define ASSET_MAP_WORLD         "assets/meta/map/overworld_tile_map.txt"
 #define ASSET_MAP_BLOCKING      "assets/meta/map/overworld_blocking_map.txt"
 
-/* États du jeu */
 typedef enum {
     STATE_MENU,
     STATE_PLAYING,
@@ -59,14 +48,12 @@ typedef enum {
     STATE_GAMEOVER
 } GameState;
 
-/* État du rendu SDL */
 typedef struct {
     SDL_Window*   window;
     SDL_Renderer* renderer;
     TTF_Font*     font;
 } RenderState;
 
-/* Statistiques du joueur */
 typedef struct {
     int score;
     int kills;
@@ -74,12 +61,11 @@ typedef struct {
     int moves;
 } PlayerStats;
 
-/* Caméra pour le mode open world */
 typedef struct {
-    int x;
-    int y;
-    int targetX;
-    int targetY;
+    float x;
+    float y;
+    float targetX;
+    float targetY;
 } Camera;
 
 #endif
